analytics service done

+++ backend/prediction-service/src/main/java/com/anghel/investmenthelper/prediction/util/property/PredictionProperties.java
@@ -14,7 +14,7 @@ public class PredictionProperties {
 
     private String modelStoragePath;
 
-    private Integer trainingRatio;
+    private Double trainingRatio;
 
     @PostConstruct
     void validate() {
